backend/requirements.txt
# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
aiosqlite==0.20.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# AI & LLM
groq==0.11.0
httpx==0.27.2
openai==1.51.0

# Vector DB & Embeddings
chromadb==0.5.5
sentence-transformers==3.1.1

# Document Processing
python-docx==1.1.2
PyPDF2==3.0.1
pdfplumber==0.11.4

# Data & Utilities
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
jinja2==3.1.4

# CORS
starlette==0.38.6

# PDF Generation
fpdf2==2.8.2
